+++ lib/screens/firstscreen/returning_List.dart
@@ -1,32 +1,32 @@
 import 'package:flutter/material.dart';
-import '/model/TodoClass.dart';
-import '/model/TodoItem.dart';
-import '/model/TodoProvider.dart';
-import 'ShowDialog.dart';
+import '/model/todo_Class.dart';
+import 'todo_Item.dart';
+import '/model/todo_Provider.dart';
+import 'show_Dialog.dart';
 
 class ReturnList extends StatefulWidget {
   final TodoProvider todoProvider;
+  final List<Todo> todos;
 
-  const ReturnList({Key? key, required this.todoProvider}) : super(key: key);
+  const ReturnList({Key? key, required this.todoProvider, required this.todos})
+      : super(key: key);
 
   @override
   State<ReturnList> createState() => _ReturnListState();
 }
 
 class _ReturnListState extends State<ReturnList> {
-  List<Todo> todos = <Todo>[];
   late TodoProvider _todoProvider;
+  late final List<Todo> _todos;
 
   Todo? _editingTodo;
-
-  //Color _selectedColor = Colors.grey;
 
   @override
   void initState() {
     super.initState();
     _todoProvider = widget.todoProvider;
-    openDatabase();
-    _loadTodos();
+    _todos = widget.todos;
+
   }
 
   Future<void> openDatabase() async {
@@ -46,20 +46,13 @@ class _ReturnListState extends State<ReturnList> {
       List<Todo> loadedTodos = await _todoProvider.getAllTodos();
 
       setState(() {
-        todos.clear();
-        todos.addAll(loadedTodos);
+        _todos.clear();
+        _todos.addAll(loadedTodos);
       });
     } catch (e) {
       print('Error loading todos: $e');
     }
   }
-
-  // Future<void> _loadTodos() async {
-  //   List<Todo> loadedTodos = await widget.todoProvider.getAllTodos();
-  //   setState(() {
-  //     todos = loadedTodos;
-  //   });
-  // }
 
   void handleTodoChange(Todo todo) {
     setState(() {
@@ -71,7 +64,7 @@ class _ReturnListState extends State<ReturnList> {
     try {
       await _todoProvider.delete(todo.id!); // Ensure id is not null
       setState(() {
-        todos.removeWhere((element) => element.id == todo.id);
+        _todos.removeWhere((element) => element.id == todo.id);
       });
     } catch (e) {
       print('Error deleting todo item: $e');
@@ -87,12 +80,12 @@ class _ReturnListState extends State<ReturnList> {
         shrinkWrap: true,
         physics: const ScrollPhysics(),
         padding: const EdgeInsets.all(8),
-        itemCount: todos.length,
+        itemCount: _todos.length,
         itemBuilder: (context, index) {
           return TodoItem(
             key: UniqueKey(),
             // Add a unique key for each TodoItem
-            todo: todos[index],
+            todo: _todos[index],
             onTodoChanged: handleTodoChange,
             removeTodo: (Todo todo) {
               // Modify the signature to accept a Todo parameter
@@ -105,9 +98,8 @@ class _ReturnListState extends State<ReturnList> {
                 builder: (BuildContext context) {
                   return DisplayAlertDialog(
                     todoProvider: _todoProvider,
-                    todo: todos[index],
-                      onTodoUpdated: _loadTodos
-                    // Pass the specific todo item to the dialog
+                    todo: _todos[index],
+                    onTodoUpdated: _loadTodos,
                   );
                 },
               );
