+++ lib/screens/firstscreen/show_Dialog.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:instagramclone/screens/first_Screen.dart';
+import 'package:instagramclone/screens/firstscreen/returning_List.dart';
 import 'dart:io';
-import '/model/TodoClass.dart';
+import '/model/todo_Class.dart';
 import 'package:image_picker/image_picker.dart';
-import '/model/TodoProvider.dart';
+import '/model/todo_Provider.dart';
 
 class DisplayAlertDialog extends StatefulWidget {
   final Todo? todo;
@@ -10,12 +12,12 @@ class DisplayAlertDialog extends StatefulWidget {
   final VoidCallback onTodoUpdated;
 
   // required this.onTodoEdit
-  const DisplayAlertDialog(
-      {Key? key,
-      required this.todoProvider,
-      this.todo,
-      required this.onTodoUpdated})
-      : super(key: key);
+  const DisplayAlertDialog({
+    Key? key,
+    required this.todoProvider,
+    this.todo,
+    required this.onTodoUpdated,
+  }) : super(key: key);
 
   @override
   _DisplayAlertDialogState createState() => _DisplayAlertDialogState();
@@ -27,12 +29,14 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
   final TextEditingController _textFieldController = TextEditingController();
   final TextEditingController _textFieldController2 = TextEditingController();
   final TextEditingController _textFieldController3 = TextEditingController();
-  String _selectedPriority = '';
+  String _selectedPriority = 'Low';
   Color _selectedColor = Colors.red[300]!;
+  int _selectedindex = 0;
 
   @override
   void initState() {
     super.initState();
+
     if (widget.todo != null) {
       _textFieldController.text = widget.todo!.title;
       _textFieldController2.text = widget.todo!.description;
@@ -48,29 +52,6 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
     });
   }
 
-  Future<void> openDatabase() async {
-    try {
-      await _todoProvider
-          .open('sample'); // Open the database with the desired name
-      print('Database opened successfully');
-    } catch (e) {
-      print('Error opening database: $e');
-    }
-  }
-
-  Future<void> _loadTodos() async {
-    print('load todos of first dispalu dialog screen');
-    try {
-      List<Todo> loadedTodos = await _todoProvider.getAllTodos();
-      setState(() {
-        // todos.clear();
-        todos.addAll(loadedTodos);
-      });
-    } catch (e) {
-      print('Error loading todos: $e');
-    }
-  }
-
   Future<void> _addOrUpdateTodoItem(Todo? todo, String title,
       String description, String text, String color) async {
     if (todo != null) {
@@ -79,7 +60,8 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
       todo.description = description;
       todo.text = text;
       todo.color = color;
-      await _todoProvider.update(todo);
+     var result= await _todoProvider.update(todo);
+      widget.onTodoUpdated();
     } else {
       // Add new todo
       Todo newTodo = Todo(
@@ -90,13 +72,25 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
         color: color,
       );
       await _todoProvider.insert(newTodo);
+
     }
-    widget.onTodoUpdated();
+  }
+
+  Future<void> _loadTodos() async {
+    print('load todos of first screen dialog');
+    try {
+      List<Todo> loadedTodos = await _todoProvider.getAllTodos();
+      setState(() {
+        //todos.clear();
+        todos.addAll(loadedTodos);
+      });
+    } catch (e) {
+      print('Error loading todos: $e');
+    }
   }
 
   Color _getPriorityColor(String color) {
-    // Implement this method to return the appropriate color based on the input string
-    // Example implementation:
+
     switch (color) {
       case 'red':
         return Colors.red;
@@ -112,19 +106,22 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
   @override
   void dispose() {
     // Dispose of the controllers to free up resources
+    print('i m dispose of show-dialog');
     _textFieldController.dispose();
     _textFieldController2.dispose();
     _textFieldController3.dispose();
+
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
+    TodoProvider todoProvider = TodoProvider();
     return SizedBox(
       width: 360.0,
       height: 400.0,
       child: AlertDialog(
-        title:  Text(
+        title: Text(
           widget.todo == null ? 'Create a Todo' : 'Update Todo',
           style: const TextStyle(
             fontWeight: FontWeight.bold,
@@ -169,7 +166,7 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
                 child: TextField(
                   controller: _textFieldController2,
                   decoration: const InputDecoration(
-                      hintText: 'Enter description'), //act as a placeholder
+                      hintText: 'Enter Description'), //act as a placeholder
                   autofocus: true,
                 ),
               ),
@@ -267,7 +264,7 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
                           elevation: _selectedPriority == 'Low' ? 20 : 5,
                           minimumSize: _selectedPriority == 'Low'
                               ? const Size.square(45)
-                              : Size.square(35),
+                              : const Size.square(35),
                           shape: const RoundedRectangleBorder(
                               borderRadius:
                                   BorderRadius.all(Radius.circular(5)))),
@@ -329,46 +326,29 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
                 ),
               ),
               onPressed: () async {
-                Navigator.of(context).pop();
+
                 String title = _textFieldController.text;
                 String description = _textFieldController2.text;
                 String text = _textFieldController3.text;
-                String color = _selectedPriority;
+                String color = getColorForPriority(_selectedPriority);
 
-                // Todo newTodo = Todo(
-                //   title: _textFieldController.text,
-                //   description: _textFieldController2.text,
-                //   completed: false,
-                //   text: _textFieldController3.text,
-                //   color: _selectedPriority,
-                // );
 
-                // if (widget.todo == null) {
-                //   print('Going to insert');
-                //   print('New Todo: $newTodo');
-                //   await widget.todoProvider.insert(newTodo);
-                // } else {
-                //   print('Going to update');
-                //   await _todoProvider.update(newTodo);
-                // }
-
-                // Clear text fields
-                // _textFieldController.clear();
-                // _textFieldController2.clear();
-                // _textFieldController3.clear();
-
-                await _loadTodos();
-                // widget.onTodoEdit();
                 await _addOrUpdateTodoItem(
                     widget.todo, title, description, text, color);
-                // Navigator.of(context).pop();
+
+
+
+                widget.onTodoUpdated();
+
+                Navigator.of(context).pop(true);
+
               },
-              child:  Padding(
+              child: Padding(
                 padding: const EdgeInsets.all(10.0),
                 child: Text(widget.todo == null ? 'Add Todo' : 'Update Todo'),
-                ),
               ),
             ),
+          ),
         ],
       ),
     );
@@ -415,5 +395,15 @@ class _DisplayAlertDialogState extends State<DisplayAlertDialog> {
         const SnackBar(content: Text('Nothing is selected')),
       );
     }
+  }
+
+  String getColorForPriority(String selectedPriority) {
+    if(selectedPriority=="High"){
+    return "Red";
+    }
+    else if(selectedPriority.contains("High"))
+      return "Orange";
+    else
+      return "Blue";
   }
 }
