+++ lib/screens/setting_Screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:instagramclone/component/BottomBar.dart';
+import 'package:instagramclone/component/bottom_Bar.dart';
 
 class Setting extends StatefulWidget {
   const Setting({Key? key}) : super(key: key);
