+++ lib/screens/start_Screen.dart
@@ -5,8 +5,6 @@ class Frontscreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-
-
     return SafeArea(
       child: Scaffold(
         backgroundColor: Colors.white,
@@ -29,7 +27,6 @@ class Frontscreen extends StatelessWidget {
                                   BorderRadius.all(Radius.circular(10)),
                             )),
                         onPressed: () {
-
                           Navigator.pushNamed(context, '/');
                           // Simulate a delay for loading
                           Future.delayed(const Duration(seconds: 1), () {
@@ -38,7 +35,7 @@ class Frontscreen extends StatelessWidget {
                         },
                         child: const Text(
                           'Skip',
-                          style:  TextStyle(
+                          style: TextStyle(
                             color: Colors.white,
                           ),
                         ),
@@ -77,7 +74,7 @@ class Frontscreen extends StatelessWidget {
                 image: AssetImage('assets/5.png'),
                 //height: 185,
               ),
-            )
+            ),
           ],
         ),
       ),
